all: Add KR_920_923 band definition
